requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.10

# Dimensionality reduction
umap-learn>=0.5.4

# Clustering
hdbscan>=0.8.33

# Config (optional)
pyyaml>=6.0
